Expose the NormalizeKey logic to the outside.
